Version 0.7.9
  1. Add RequestFilter and DeserializeFilter for JWT token web request.
  2. Add enable_html_description and enable_html_short_description flags for WooCommerce Product.
  3. Add Debug mode for RestAPI. NOTE: Beware when setting Debug to True, as exceptions might contain sensetive information.
  4. Change total_sales from int (Int32) to long (Int64).
  5. Fix OrderNote can't be force deleted bug.

+++ WooCommerce/Legacy/Product.cs
@@ -413,7 +413,7 @@ namespace WooCommerceNET.WooCommerce.Legacy
         /// read-only
         /// </summary>
         [DataMember(EmitDefaultValue = false)]
-        public int? total_sales { get; set; }
+        public long? total_sales { get; set; }
 
         /// <summary>
         /// List of products variations. See Variations Properties
